feat: 도메인 예외 검증 추가

+++ src/main/java/cart/application/CartItemService.java
@@ -9,7 +9,6 @@ import cart.dto.CartItemQuantityUpdateRequest;
 import cart.dto.CartItemRequest;
 import cart.dto.CartItemResponse;
 import cart.exception.CartItemException;
-import cart.exception.ProductNotExistsException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -35,19 +34,21 @@ public class CartItemService {
         Product product = productDao.getProductById(cartItemRequest.getProductId());
 
         if (Objects.isNull(product)) {
-            throw new ProductNotExistsException("카트에 추가하려는 상품이 존재하지 않습니다.");
+            throw new CartItemException.NotFound("카트에 추가하려는 상품이 존재하지 않습니다.");
         }
 
         return cartItemDao.save(new CartItem(member, product));
     }
 
     public void updateQuantity(Member member, Long id, CartItemQuantityUpdateRequest request) {
-
+        if (Objects.isNull(id)) {
+            throw new CartItemException("장바구니 아이디를 입력해야 합니다.");
+        }
 
         CartItem cartItem = cartItemDao.findById(id);
 
         if (Objects.isNull(cartItem)) {
-            throw new CartItemException("존재하지 않는 장바구니 아이템입니다.");
+            throw new CartItemException("존재하지 않는 장바구니 상품입니다.");
         }
 
         cartItem.checkOwner(member);
@@ -62,7 +63,16 @@ public class CartItemService {
     }
 
     public void remove(Member member, Long id) {
+        if (Objects.isNull(id)) {
+            throw new CartItemException("장바구니 아이디를 입력해야 합니다");
+        }
+
         CartItem cartItem = cartItemDao.findById(id);
+
+        if (Objects.isNull(cartItem)) {
+            throw new CartItemException.CartItemNotExists("존재하지 않는 장바구니 상품입니다.");
+        }
+
         cartItem.checkOwner(member);
 
         cartItemDao.deleteById(id);
